+++ src/nami_sdk/__init__.py
@@ -1,5 +1,6 @@
 """Nami Core Python SDK."""
 
 from nami_sdk.client import NamiClient
+from nami_sdk.async_client import NamiAsyncClient, NamiWSListener
 
-__all__ = ["NamiClient"]
+__all__ = ["NamiClient", "NamiAsyncClient", "NamiWSListener"]